refactor: move database to package
