+++ app/src/main/java/com/madrid/movio/di/DataModule.kt
@@ -1,7 +1,7 @@
 package com.madrid.movio.di
 
 import com.madrid.data.dataSource.local.LocalDataSourceImpl
-import com.madrid.data.dataSource.local.MovioDatabase
+import com.madrid.data.dataSource.local.database.MovioDatabase
 import com.madrid.data.repositories.AllTrendingRepositoryImpl
 import com.madrid.data.repositories.ArtistDetailsRepositoryImpl
 import com.madrid.data.repositories.HomeRepositoryImpl
